feat: add ability to add default values
In order to add a new default you need to follow the following steps:
Add the new hierarchy and default value to Profiles.json. In the example there, the hierarchy is "/zgw/zaken/zaken" (based on OpenApi spec) and the element to have a default value is betalingsindicatie.

Next you have to add your new hierarchy (which in the example is 'zgw/zaken/zaken') to ExtractProfileDefaults.xsl if it does not already exist there.

Lastly you must add a sender that calls ImportProfileDefaults with a param named "hierarchy" with the value of your hierarchy. ('zgw/zaken/zaken' as shown above.)